#11509 - The deletion of actions and campaign forms from another jurisdiction should not be allowed

+++ sormas-backend/src/main/java/de/symeda/sormas/backend/campaign/data/CampaignFormDataFacadeEjb.java
@@ -58,6 +58,7 @@ import de.symeda.sormas.api.campaign.diagram.CampaignDiagramSeries;
 import de.symeda.sormas.api.campaign.form.CampaignFormElement;
 import de.symeda.sormas.api.campaign.form.CampaignFormElementType;
 import de.symeda.sormas.api.i18n.I18nProperties;
+import de.symeda.sormas.api.i18n.Strings;
 import de.symeda.sormas.api.i18n.Validations;
 import de.symeda.sormas.api.infrastructure.PopulationDataCriteria;
 import de.symeda.sormas.api.infrastructure.PopulationDataDto;
@@ -65,6 +66,7 @@ import de.symeda.sormas.api.infrastructure.area.AreaReferenceDto;
 import de.symeda.sormas.api.infrastructure.district.DistrictReferenceDto;
 import de.symeda.sormas.api.infrastructure.region.RegionReferenceDto;
 import de.symeda.sormas.api.user.UserRight;
+import de.symeda.sormas.api.utils.AccessDeniedException;
 import de.symeda.sormas.api.utils.SortProperty;
 import de.symeda.sormas.api.utils.ValidationRuntimeException;
 import de.symeda.sormas.backend.campaign.Campaign;
@@ -210,6 +212,11 @@ public class CampaignFormDataFacadeEjb implements CampaignFormDataFacade {
 		}
 
 		CampaignFormData campaignFormData = campaignFormDataService.getByUuid(campaignFormDataUuid);
+
+		if (!campaignFormDataService.inJurisdictionOrOwned(campaignFormData)) {
+			throw new AccessDeniedException(I18nProperties.getString(Strings.messageCampaignFormOutsideJurisdictionDeletionDenied));
+		}
+
 		campaignFormDataService.deletePermanent(campaignFormData);
 	}
 
@@ -404,36 +411,36 @@ public class CampaignFormDataFacadeEjb implements CampaignFormDataFacade {
 //						diagramSeries.getFormId(),
 //						false));
 //			} else {
-				regions.stream().forEach(regionReferenceDto -> {
-					PopulationDataCriteria criteria = new PopulationDataCriteria();
-					criteria.sexIsNull(true);
-					criteria.region(regionReferenceDto);
-					criteria.ageGroup(diagramSeriesTotal.getPopulationGroup());
-					List<PopulationDataDto> populationDataDto = populationDataFacadeEjb.getPopulationData(criteria);
-					Integer populationSum = 0;
-					if (!populationDataDto.isEmpty()) {
-						populationSum = populationDataDto.stream().mapToInt(e -> e.getPopulation()).sum();
-						resultData.add(
-							new CampaignDiagramDataDto(
-								regionReferenceDto.getCaption(),
-								populationSum,
-								regionReferenceDto.getUuid(),
-								regionReferenceDto.getCaption(),
-								diagramSeries.getFieldId(),
-								diagramSeries.getFormId(),
-								true));
-					} else {
-						resultData.add(
-							new CampaignDiagramDataDto(
-								regionReferenceDto.getCaption(),
-								0,
-								regionReferenceDto.getUuid(),
-								regionReferenceDto.getCaption(),
-								diagramSeries.getFieldId(),
-								diagramSeries.getFormId(),
-								false));
-					}
-				});
+			regions.stream().forEach(regionReferenceDto -> {
+				PopulationDataCriteria criteria = new PopulationDataCriteria();
+				criteria.sexIsNull(true);
+				criteria.region(regionReferenceDto);
+				criteria.ageGroup(diagramSeriesTotal.getPopulationGroup());
+				List<PopulationDataDto> populationDataDto = populationDataFacadeEjb.getPopulationData(criteria);
+				Integer populationSum = 0;
+				if (!populationDataDto.isEmpty()) {
+					populationSum = populationDataDto.stream().mapToInt(e -> e.getPopulation()).sum();
+					resultData.add(
+						new CampaignDiagramDataDto(
+							regionReferenceDto.getCaption(),
+							populationSum,
+							regionReferenceDto.getUuid(),
+							regionReferenceDto.getCaption(),
+							diagramSeries.getFieldId(),
+							diagramSeries.getFormId(),
+							true));
+				} else {
+					resultData.add(
+						new CampaignDiagramDataDto(
+							regionReferenceDto.getCaption(),
+							0,
+							regionReferenceDto.getUuid(),
+							regionReferenceDto.getCaption(),
+							diagramSeries.getFieldId(),
+							diagramSeries.getFormId(),
+							false));
+				}
+			});
 //			}
 		} else if (grouping == CampaignJurisdictionLevel.DISTRICT || Objects.isNull(district)) {
 
@@ -458,37 +465,37 @@ public class CampaignFormDataFacadeEjb implements CampaignFormDataFacade {
 //						diagramSeries.getFormId(),
 //						false));
 //			} else {
-				districts.stream().forEach(districtReferenceDto -> {
-					PopulationDataCriteria criteria = new PopulationDataCriteria();
-					criteria.sexIsNull(true);
-					criteria.district(districtReferenceDto);
-					criteria.region(region);
-					criteria.ageGroup(diagramSeriesTotal.getPopulationGroup());
-					List<PopulationDataDto> populationDataDtoList = populationDataFacadeEjb.getPopulationData(criteria);
-					Integer populationSum = 0;
-					if (!populationDataDtoList.isEmpty()) {
-						populationSum = populationDataDtoList.stream().mapToInt(e -> e.getPopulation()).sum();
-						resultData.add(
-							new CampaignDiagramDataDto(
-								districtReferenceDto.getCaption(),
-								populationSum,
-								districtReferenceDto.getUuid(),
-								districtReferenceDto.getCaption(),
-								diagramSeries.getFieldId(),
-								diagramSeries.getFormId(),
-								true));
-					} else {
-						resultData.add(
-							new CampaignDiagramDataDto(
-								districtReferenceDto.getCaption(),
-								populationSum,
-								districtReferenceDto.getUuid(),
-								districtReferenceDto.getCaption(),
-								diagramSeries.getFieldId(),
-								diagramSeries.getFormId(),
-								false));
-					}
-				});
+			districts.stream().forEach(districtReferenceDto -> {
+				PopulationDataCriteria criteria = new PopulationDataCriteria();
+				criteria.sexIsNull(true);
+				criteria.district(districtReferenceDto);
+				criteria.region(region);
+				criteria.ageGroup(diagramSeriesTotal.getPopulationGroup());
+				List<PopulationDataDto> populationDataDtoList = populationDataFacadeEjb.getPopulationData(criteria);
+				Integer populationSum = 0;
+				if (!populationDataDtoList.isEmpty()) {
+					populationSum = populationDataDtoList.stream().mapToInt(e -> e.getPopulation()).sum();
+					resultData.add(
+						new CampaignDiagramDataDto(
+							districtReferenceDto.getCaption(),
+							populationSum,
+							districtReferenceDto.getUuid(),
+							districtReferenceDto.getCaption(),
+							diagramSeries.getFieldId(),
+							diagramSeries.getFormId(),
+							true));
+				} else {
+					resultData.add(
+						new CampaignDiagramDataDto(
+							districtReferenceDto.getCaption(),
+							populationSum,
+							districtReferenceDto.getUuid(),
+							districtReferenceDto.getCaption(),
+							diagramSeries.getFieldId(),
+							diagramSeries.getFormId(),
+							false));
+				}
+			});
 //			}
 		} else if (district != null) {
 			resultData.add(
@@ -627,8 +634,10 @@ public class CampaignFormDataFacadeEjb implements CampaignFormDataFacade {
 			}
 
 			// WHERE
-			StringBuilder whereBuilder =
-				new StringBuilder(" WHERE ").append(CampaignFormMeta.TABLE_NAME).append(".").append(CampaignFormMeta.FORM_ID).append(" = :campaignFormMetaId");
+			StringBuilder whereBuilder = new StringBuilder(" WHERE ").append(CampaignFormMeta.TABLE_NAME)
+				.append(".")
+				.append(CampaignFormMeta.FORM_ID)
+				.append(" = :campaignFormMetaId");
 
 			if (series.getFieldId() != null) {
 				whereBuilder.append(" AND jsonData->>'")
